+++ rosariosis.sql
@@ -3559,7 +3559,7 @@ INSERT INTO attendance_codes VALUES (4, 2017, 1, 'Excused Absence', 'E', 'offici
 --
 
 INSERT INTO config VALUES (0, 'LOGIN', 'No');
-INSERT INTO config VALUES (0, 'VERSION', '3.4.3');
+INSERT INTO config VALUES (0, 'VERSION', '3.');
 INSERT INTO config VALUES (0, 'TITLE', 'Rosario Student Information System');
 INSERT INTO config VALUES (0, 'NAME', 'RosarioSIS');
 INSERT INTO config VALUES (0, 'MODULES', 'a:13:{s:12:"School_Setup";b:1;s:8:"Students";b:1;s:5:"Users";b:1;s:10:"Scheduling";b:1;s:6:"Grades";b:1;s:10:"Attendance";b:1;s:11:"Eligibility";b:1;s:10:"Discipline";b:1;s:10:"Accounting";b:1;s:15:"Student_Billing";b:1;s:12:"Food_Service";b:1;s:9:"Resources";b:1;s:6:"Custom";b:1;}');
@@ -3570,6 +3570,7 @@ INSERT INTO config VALUES (0, 'CREATE_USER_ACCOUNT', NULL);
 INSERT INTO config VALUES (0, 'CREATE_STUDENT_ACCOUNT', NULL);
 INSERT INTO config VALUES (0, 'STUDENTS_EMAIL_FIELD', NULL);
 INSERT INTO config VALUES (0, 'LIMIT_EXISTING_CONTACTS_ADDRESSES', NULL);
+INSERT INTO config VALUES (0, 'FAILED_LOGIN_LIMIT', NULL);
 INSERT INTO config VALUES (1, 'SCHOOL_SYEAR_OVER_2_YEARS', 'Y');
 INSERT INTO config VALUES (1, 'ATTENDANCE_FULL_DAY_MINUTES', '300');
 INSERT INTO config VALUES (1, 'STUDENTS_USE_MAILING', NULL);